Move QueryTest -> harness.py
